requirements.txt
asgiref==3.3.1
dj-static==0.0.6
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.12.2
djongo==1.3.3
dnspython==2.0.0
gunicorn==20.0.4
pymongo==3.11.0
pytz==2020.4
sqlparse==0.2.4
static3==0.7.0
